refactor(getmoderstats): Refactors moderation stats system and fixes type inconsistencies
Improves moderation statistics handling by extracting business logic into dedicated utility modules and introducing proper type annotations for infractions data.

Changes ChangeStat amount field from integer to float to support decimal scoring values.

Consolidates moderation stats calculations into reusable functions and separates presentation logic from data retrieval. Adds TYPE_CHECKING guards to prevent circular imports.

Updates infractions view to display total punishment count and enhances UI with improved formatting and color theming.

Re-enables previously disabled background tasks for punishment expiration, ticket deletion, and clan reputation management.

Fixes task scheduling to use non-blocking message sends via asyncio.create_task to prevent task delays.

+++ src/infra/db/models/changestat.py
@@ -13,7 +13,7 @@ from src.infra.db.models.base import Base
 class ChangeStat(IdIntegerMixin, Base):
     guild_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
     moderator_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
-    amount: Mapped[int] = mapped_column(nullable=False)
+    amount: Mapped[float] = mapped_column(nullable=False)
     reason: Mapped[str] = mapped_column(String, nullable=False)
     type: Mapped[ChangeStatTypeEnum] = mapped_column(
         Enum(
